+++ CHANGES.rst
@@ -6,6 +6,10 @@ PRAW follows `semantic versioning <http://semver.org/>`_.
 Unreleased
 ----------
 
+**Added**
+
+- :meth:`~.SubredditLinkFlairTemplates.user_selectable` to get available subreddit link flairs
+
 7.4.0 (2021/07/30)
 ------------------